+++ src/main/java/com/urbanfleet/service/VehicleService.java
@@ -1,13 +1,17 @@
 package com.urbanfleet.service;
 
 import com.urbanfleet.dto.VehicleRequest;
+import com.urbanfleet.dto.VehicleResponse;
 import com.urbanfleet.model.Vehicle;
+import com.urbanfleet.repository.VehicleRepository;
+import org.springframework.beans.factory.annotation.Autowired;
 
 import java.util.List;
 
 public interface VehicleService
 {
+
     String savevehicle(VehicleRequest request);
-    List<Vehicle> getAllVehicleData();
+    public List<VehicleResponse> getAllVehicleData();
 
 }